--- apps/roboshield/src/components/RoboForm/index.ts
@@ -1,3 +1,0 @@
-import RoboForm from "./RoboForm";
-
-export default RoboForm;
