+++ apps/roboshield/src/components/RobotsGenerator/index.ts
@@ -1,0 +1,3 @@
+import RobotsGenerator from "./RobotsGenerator";
+
+export default RobotsGenerator;
